test(e2e): address review — share findCollectionDir, robust folder-settings frame
- extract findCollectionDir into tests/e2e/utils/page/actions.ts (was copy-pasted)
- add data-testid="folder-settings" and key the folder-settings frame on it,
  instead of the ambiguous [role="tab"] scan that could match a request editor
- drop task-bound AC-number labels from spec comments (ai-hygiene)
